- Util file added for fetching installations and connections from Github - Updated `jira-get` and the new `subscriptions-get` route to use the util methods
