+++ src/rest/routes/subscriptions/index.tsx
@@ -4,9 +4,9 @@ import { getAllSubscriptions } from "./service";
 
 export const SubscriptionsRouter = Router();
 
-SubscriptionsRouter.get("/", errorWrapper("SubscriptionsGet", async (_req: Request, res: Response) => {
+SubscriptionsRouter.get("/", errorWrapper("SubscriptionsGet", async (req: Request, res: Response) => {
 	const { jiraHost, installation } = res.locals;
-	const { ghCloudSubscriptions, ghEnterpriseServers } = await getAllSubscriptions(jiraHost, installation.id);
+	const { ghCloudSubscriptions, ghEnterpriseServers } = await getAllSubscriptions(jiraHost, installation.id, req);
 
 	res.status(200).json({
 		ghCloudSubscriptions,
